+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.0.0
+
+* Added semantic versioning.
+* `call` renamed to `apply`.
+* New `call` works like `apply` but accepts parameters for server method as a list of function arguments.
+* `subid` property of the subscription object renamed to `subscriptionId`.
+* Added `subscribe` method. Works like `sub` but accepts parameters for server publication as a list of function arguments.
+
 # 1.2.3
 
 * Updated `simpleddp-core` package.